+++ common/module.js
@@ -1,1 +1,3 @@
+'use strict';
+
 angular.module('frapontillo.bootstrap-switch', []);