+++ example/scripts/controllers/main.js
@@ -8,10 +8,10 @@ angular.module('bsSwitchApp')
     $scope.isActive = true;
     $scope.size = 'normal';
     $scope.animate = true;
-    $scope.radioOff = false;
-    $scope.icon = 'icon-youtube';
+    $scope.radioOff = true;
     $scope.handleWidth = "auto";
     $scope.labelWidth = "auto";
+    $scope.inverse = true;
 
     $scope.$watch('isSelected', function() {
       $log.info('Selection changed.');