+++ src/directives/bsSwitch.js
@@ -13,7 +13,10 @@ angular.module('frapontillo.bootstrap-switch')
          * @returns {Object} representing the true view value; if undefined, returns true.
          */
         var getTrueValue = function() {
-          var trueValue = $parse(attrs.ngTrueValue)(scope);
+          if (attrs.type === 'radio') {
+            return attrs.value || $parse(attrs.ngValue)(scope) || true;
+          }
+          var trueValue = ($parse(attrs.ngTrueValue)(scope));
           if (!angular.isString(trueValue)) {
             trueValue = true;
           }
@@ -21,11 +24,20 @@ angular.module('frapontillo.bootstrap-switch')
         };
 
         /**
-         * Get a boolean value from a boolean-like string.
+         * Get a boolean value from a boolean-like string, evaluating it on the current scope.
          * @param value The input object
          * @returns {boolean} A boolean value
          */
         var getBooleanFromString = function(value) {
+          return scope.$eval(value) === true;
+        };
+
+        /**
+         * Get a boolean value from a boolean-like string, defaulting to true if undefined.
+         * @param value The input object
+         * @returns {boolean} A boolean value
+         */
+        var getBooleanFromStringDefTrue = function(value) {
           return (value === true || value === 'true' || !value);
         };
 
@@ -48,15 +60,11 @@ angular.module('frapontillo.bootstrap-switch')
          */
         var getSwitchAttrValue = function(attrName) {
           var map = {
-            'switchRadioOff': function(value) {
-              return (value === true || value === 'true');
-            },
+            'switchRadioOff': getBooleanFromStringDefTrue,
             'switchActive': function(value) {
-              return !getBooleanFromString(value);
+              return !getBooleanFromStringDefTrue(value);
             },
-            'switchAnimate': function(value) {
-              return scope.$eval(value || 'true');
-            },
+            'switchAnimate': getBooleanFromStringDefTrue,
             'switchLabel': function(value) {
               return value ? value : '&nbsp;';
             },
@@ -67,7 +75,9 @@ angular.module('frapontillo.bootstrap-switch')
             },
             'switchWrapper': function(value) {
               return value || 'wrapper';
-            }
+            },
+            'switchInverse': getBooleanFromString,
+            'switchReadonly': getBooleanFromString
           };
           var transFn = map[attrName] || getValueOrUndefined;
           return transFn(attrs[attrName]);
@@ -116,9 +126,15 @@ angular.module('frapontillo.bootstrap-switch')
               labelText: attrs.switchLabel ? getSwitchAttrValue('switchLabel') : getSwitchAttrValue('switchIcon'),
               wrapperClass: getSwitchAttrValue('switchWrapper'),
               handleWidth: getSwitchAttrValue('switchHandleWidth'),
-              labelWidth: getSwitchAttrValue('switchLabelWidth')
+              labelWidth: getSwitchAttrValue('switchLabelWidth'),
+              inverse: getSwitchAttrValue('switchInverse'),
+              readonly: getSwitchAttrValue('switchReadonly')
             });
-            controller.$setViewValue(viewValue);
+            if (attrs.type === 'radio') {
+              controller.$setViewValue(controller.$modelValue);
+            } else {
+              controller.$setViewValue(viewValue);
+            }
           }
         };
 
@@ -128,7 +144,7 @@ angular.module('frapontillo.bootstrap-switch')
         var listenToModel = function () {
 
           attrs.$observe('switchActive', function (newValue) {
-            var active = getBooleanFromString(newValue);
+            var active = getBooleanFromStringDefTrue(newValue);
             // if we are disabling the switch, delay the deactivation so that the toggle can be switched
             if (!active) {
               $timeout(function() {
@@ -140,11 +156,15 @@ angular.module('frapontillo.bootstrap-switch')
             }
           });
 
+          function modelValue() {
+            return controller.$modelValue;
+          }
+
           // When the model changes
-          scope.$watch(attrs.ngModel, function(newValue) {
+          scope.$watch(modelValue, function(newValue) {
             initMaybe();
             if (newValue !== undefined) {
-              element.bootstrapSwitch('state', newValue === getTrueValue(), true);
+              element.bootstrapSwitch('state', newValue === getTrueValue(), false);
             }
           }, true);
 
@@ -161,7 +181,9 @@ angular.module('frapontillo.bootstrap-switch')
             'switchIcon': 'labelText',
             'switchWrapper': 'wrapperClass',
             'switchHandleWidth': 'handleWidth',
-            'switchLabelWidth': 'labelWidth'
+            'switchLabelWidth': 'labelWidth',
+            'switchInverse': 'inverse',
+            'switchReadonly': 'readonly'
           };
 
           var observeProp = function(prop, bindings) {
@@ -182,11 +204,29 @@ angular.module('frapontillo.bootstrap-switch')
          * Listen to view changes.
          */
         var listenToView = function () {
-          // When the switch is clicked, set its value into the ngModel
-          element.on('switchChange.bootstrapSwitch', function (e, data) {
-            // $setViewValue --> $viewValue --> $parsers --> $modelValue
-            controller.$setViewValue(data);
-          });
+          if (attrs.type === 'radio') {
+            // when the switch is clicked
+            element.on('change.bootstrapSwitch', function (e) {
+              // discard not real change events
+              if ((controller.$modelValue === controller.$viewValue) && (e.target.checked !== $(e.target).bootstrapSwitch('state'))) {
+                // $setViewValue --> $viewValue --> $parsers --> $modelValue
+                // if the switch is indeed selected
+                if (e.target.checked) {
+                  // set its value into the view
+                  controller.$setViewValue(getTrueValue());
+                } else if (getTrueValue() === controller.$viewValue) {
+                  // otherwise if it's been deselected, delete the view value
+                  controller.$setViewValue(undefined);
+                }
+              }
+            });
+          } else {
+            // When the checkbox switch is clicked, set its value into the ngModel
+            element.on('switchChange.bootstrapSwitch', function (e) {
+              // $setViewValue --> $viewValue --> $parsers --> $modelValue
+              controller.$setViewValue(e.target.checked);
+            });
+          }
         };
 
         // Listen and respond to view changes