+++ test/spec/directives/bsSwitchSpec.js
@@ -24,13 +24,21 @@ describe('Directive: bsSwitch', function () {
       scope: {model:true},
       element: 'ng-model="model" type="checkbox"'
     },
+    'multipleRadios': {
+      scope: {model:''},
+      element: [
+        'ng-model="model" name="radio" type="radio" value="uno"',
+        'ng-model="model" name="radio" type="radio" value="dos"',
+        'ng-model="model" name="radio" type="radio" value="tres"'
+      ]
+    },
     'radio': {
       scope: {model:true},
-      element: 'ng-model="model" type="radio"'
+      element: 'ng-model="model" name="radio" type="radio"'
     },
     'radioOff': {
       scope: {model:true, radioOff:false},
-      element: 'ng-model="model" type="radio" switch-radio-off="{{ radioOff }}"'
+      element: 'ng-model="model" name="radio" type="radio" switch-radio-off="{{ radioOff }}"'
     },
     'active': {
       scope: {model:true, isActive:true},
@@ -39,6 +47,10 @@ describe('Directive: bsSwitch', function () {
     'unactivated': {
       scope: {model:true, isActive:false},
       element: 'ng-model="model" type="checkbox" switch-active="{{ isActive }}"'
+    },
+    'readonly': {
+      scope: {model:true},
+      element: 'ng-model="model" type="checkbox" switch-readonly="{{ isReadonly }}"'
     },
     'size': {
       scope: {model:true, size:'large'},
@@ -67,15 +79,26 @@ describe('Directive: bsSwitch', function () {
     'customValues': {
       scope: {model:'something'},
       element: 'ng-model="model" type="checkbox" ng-true-value="\'yep\'" ng-false-value="\'nope\'"'
+    },
+    'inverse': {
+      scope: {model:true},
+      element: 'ng-model="model" type="checkbox" switch-inverse="{{ inverse }}"'
+    },
+    'getterSetter': {
+      scope: {},
+      element: 'ng-model="modelGetterSetter" ng-model-options="{getterSetter: true}" type="checkbox"'
     }
   };
 
   var CONST = {
     SWITCH_CLASS: 'bootstrap-switch',
     SWITCH_WRAPPER_CLASS: 'bootstrap-switch-wrapper',
+    SWITCH_CONTAINER_CLASS: 'bootstrap-switch-container',
+    SWITCH_INVERSE_CLASS: 'bootstrap-switch-inverse',
     SWITCH_ON_CLASS: 'bootstrap-switch-on',
     SWITCH_OFF_CLASS: 'bootstrap-switch-off',
     SWITCH_DISABLED_CLASS: 'bootstrap-switch-disabled',
+    SWITCH_READONLY_CLASS: 'bootstrap-switch-readonly',
     SWITCH_MINI_CLASS: 'bootstrap-switch-mini',
     SWITCH_INFO_CLASS: 'bootstrap-switch-info',
     SWITCH_WARNING_CLASS: 'bootstrap-switch-warning',
@@ -99,11 +122,24 @@ describe('Directive: bsSwitch', function () {
    */
   function buildElement(template, input) {
     var elementContent = template.element;
-    var realElement = (input ? '<input ' : '<') + 'bs-switch ' + elementContent + '>';
-    if (!input) {
-      realElement += '</bs-switch>';
+    var realElement;
+    if (angular.isArray(elementContent)) {
+      realElement = '<div>';
+      for (var c in elementContent) {
+        realElement += buildSingleElement(elementContent[c], input);
+      }
+      realElement += '</div>';
+      return realElement;
     }
-    return realElement;
+    return buildSingleElement(elementContent, input);
+  }
+
+  function buildSingleElement(content, isInput) {
+    var singleElement = (isInput ? '<input ' : '<') + 'bs-switch ' + content + '>';
+    if (!isInput) {
+      singleElement += '</bs-switch>';
+    }
+    return singleElement;
   }
 
   /**
@@ -181,6 +217,62 @@ describe('Directive: bsSwitch', function () {
   it('should change a radio from true to false', inject(makeTestRadioOffTrue()));
   it('should change a radio from true to false (input)', inject(makeTestRadioOffTrue(true)));
 
+
+  function expectNothing(el1, el2, el3) {
+    expect(el1.hasClass(CONST.SWITCH_OFF_CLASS)).toBeTruthy();
+    expect(el1.hasClass(CONST.SWITCH_ON_CLASS)).toBeFalsy();
+    expect(el2.hasClass(CONST.SWITCH_OFF_CLASS)).toBeTruthy();
+    expect(el2.hasClass(CONST.SWITCH_ON_CLASS)).toBeFalsy();
+    expect(el3.hasClass(CONST.SWITCH_OFF_CLASS)).toBeTruthy();
+    expect(el3.hasClass(CONST.SWITCH_ON_CLASS)).toBeFalsy();
+  }
+
+  function makeTestMultipleRadios(input) {
+    return function () {
+      var element = compileDirective('multipleRadios', input);
+      var elements = element.find('.bootstrap-switch');
+      var el1 = angular.element(elements[0]);
+      var el2 = angular.element(elements[1]);
+      var el3 = angular.element(elements[2]);
+      expectNothing(el1, el2, el3);
+      scope.model = 'wat';
+      scope.$apply();
+      expectNothing(el1, el2, el3);
+      scope.model = 'dos';
+      scope.$apply();
+      expect(el2.hasClass(CONST.SWITCH_OFF_CLASS)).toBeFalsy();
+      expect(el2.hasClass(CONST.SWITCH_ON_CLASS)).toBeTruthy();
+      expect(scope.model).toEqual('dos');
+      expect(el1.hasClass(CONST.SWITCH_OFF_CLASS)).toBeTruthy();
+      expect(el1.hasClass(CONST.SWITCH_ON_CLASS)).toBeFalsy();
+      expect(el3.hasClass(CONST.SWITCH_OFF_CLASS)).toBeTruthy();
+      expect(el3.hasClass(CONST.SWITCH_ON_CLASS)).toBeFalsy();
+    };
+  }
+  it('should set the proper model with multiple radios', inject(makeTestMultipleRadios()));
+  it('should set the proper model with multiple radios (input)', inject(makeTestMultipleRadios(true)));
+
+  function makeTestMultipleRadiosOff(input) {
+    return function () {
+      var element = compileDirective('multipleRadios', input);
+      var elements = element.find('.bootstrap-switch');
+      expect(scope.model).toEqual('');
+      var el1 = angular.element(elements[0]);
+      var el2 = angular.element(elements[1]);
+      var el3 = angular.element(elements[2]);
+      expectNothing(el1, el2, el3);
+      jQuery(el3).find('input').bootstrapSwitch('toggleState');
+      scope.$apply();
+      expect(scope.model).toEqual('tres');
+      jQuery(el3).find('input').bootstrapSwitch('toggleState');
+      scope.$apply();
+      expect(scope.model).toEqual(undefined);
+      expectNothing(el1, el2, el3);
+    };
+  }
+  it('should set the proper model to undefined when a radio is turned off', inject(makeTestMultipleRadiosOff()));
+  it('should set the proper model to undefined when a radio is turned off (input)', inject(makeTestMultipleRadiosOff(true)));
+
   // Test the model change
   function makeTestChangeModel(input) {
     return function () {
@@ -201,12 +293,10 @@ describe('Directive: bsSwitch', function () {
     return function () {
       var element = compileDirective(undefined, input);
       expect(scope.model).toBeTruthy();
-      // The click on the element's label executes asynchronously,
-      // so we skip that and rely on the fact that the click calls:
-      element.find(CONST.SWITCH_LEFT_SELECTOR).trigger('click.bootstrapSwitch');
+      element.find('input').bootstrapSwitch('toggleState');
       scope.$apply();
       expect(scope.model).toBeFalsy();
-      element.find(CONST.SWITCH_RIGHT_SELECTOR).trigger('click.bootstrapSwitch');
+      element.find('input').bootstrapSwitch('toggleState');
       scope.$apply();
       expect(scope.model).toBeTruthy();
     };
@@ -270,6 +360,21 @@ describe('Directive: bsSwitch', function () {
   }
   it('should activate the switch', inject(makeTestActivate()));
   it('should activate the switch (input)', inject(makeTestActivate(true)));
+
+  // Test the readonly
+  function makeTestReadonly(input) {
+    return function () {
+      var element = compileDirective('readonly', input);
+      expect(element.hasClass(CONST.SWITCH_READONLY_CLASS)).toBeFalsy();
+      expect(element.find(CONST.INPUT_SELECTOR).attr('readonly')).toBeFalsy();
+      scope.isReadonly = true;
+      scope.$apply();
+      expect(element.hasClass(CONST.SWITCH_READONLY_CLASS)).toBeTruthy();
+      expect(element.find(CONST.INPUT_SELECTOR).attr('readonly')).toBeTruthy();
+    };
+  }
+  it('should set the switch as read only', inject(makeTestReadonly()));
+  it('should set the switch as read only (input)', inject(makeTestReadonly(true)));
 
   // Test the size change
   function makeTestChangeSize(input) {
@@ -370,14 +475,19 @@ describe('Directive: bsSwitch', function () {
   // Test the animation deactivation and reactivation
   function makeTestAnimation(input) {
     return function () {
+      jasmine.clock().install();
       var element = compileDirective('animation', input);
+      jasmine.clock().tick(50);
       expect(element.hasClass(CONST.SWITCH_ANIMATED_CLASS)).toBeTruthy();
       scope.animate = false;
       scope.$apply();
+      jasmine.clock().tick(50);
       expect(element.hasClass(CONST.SWITCH_ANIMATED_CLASS)).toBeFalsy();
       scope.animate = true;
       scope.$apply();
+      jasmine.clock().tick(50);
       expect(element.hasClass(CONST.SWITCH_ANIMATED_CLASS)).toBeTruthy();
+      jasmine.clock().uninstall();
     };
   }
   it('should change the switch animation mode', inject(makeTestAnimation()));
@@ -442,5 +552,68 @@ describe('Directive: bsSwitch', function () {
   }
   it('should use "yep" and "nope" instead of true and false', inject(makeTestCustomValues()));
   it('should use "yep" and "nope" instead of true and false (input)', inject(makeTestCustomValues(true)));
+
+  // Test the inverse default option
+  function makeTestInverseUndefined(input) {
+    return function () {
+      var element = compileDirective('inverse', input);
+      expect(element.hasClass(CONST.SWITCH_INVERSE_CLASS)).toBeFalsy();
+      var children = element.find('.' + CONST.SWITCH_CONTAINER_CLASS).find('*[class^=\'bootstrap-switch-handle-\']');
+      expect(children.length).toBe(2);
+      expect(angular.element(children[0]).hasClass(CONST.SWITCH_LEFT_SELECTOR.substr(1))).toBeTruthy();
+      expect(angular.element(children[1]).hasClass(CONST.SWITCH_RIGHT_SELECTOR.substr(1))).toBeTruthy();
+    };
+  }
+  it('should default to inverse false when not defined', inject(makeTestInverseUndefined()));
+  it('should default to inverse false when not defined (input)', inject(makeTestInverseUndefined(true)));
+
+  // Test the inverse option
+  function makeTestInverse(input) {
+    return function () {
+      var element = compileDirective('inverse', input);
+      expect(element.hasClass(CONST.SWITCH_INVERSE_CLASS)).toBeFalsy();
+      // invert
+      scope.inverse = true;
+      scope.$apply();
+      expect(element.hasClass(CONST.SWITCH_INVERSE_CLASS)).toBeTruthy();
+      var children = element.find('.' + CONST.SWITCH_CONTAINER_CLASS).find('*[class^=\'bootstrap-switch-handle-\']');
+      expect(children.length).toBe(2);
+      expect(angular.element(children[1]).hasClass(CONST.SWITCH_LEFT_SELECTOR.substr(1))).toBeTruthy();
+      expect(angular.element(children[0]).hasClass(CONST.SWITCH_RIGHT_SELECTOR.substr(1))).toBeTruthy();
+      // reset
+      scope.inverse = false;
+      scope.$apply();
+      expect(element.hasClass(CONST.SWITCH_INVERSE_CLASS)).toBeFalsy();
+      children = element.find('.' + CONST.SWITCH_CONTAINER_CLASS).find('*[class^=\'bootstrap-switch-handle-\']');
+      expect(children.length).toBe(2);
+      expect(angular.element(children[0]).hasClass(CONST.SWITCH_LEFT_SELECTOR.substr(1))).toBeTruthy();
+      expect(angular.element(children[1]).hasClass(CONST.SWITCH_RIGHT_SELECTOR.substr(1))).toBeTruthy();
+    };
+  }
+  it('should invert the on and off switches and then reset them', inject(makeTestInverse()));
+  it('should invert the on and off switches and then reset them (input)', inject(makeTestInverse(true)));
+
+  // Test the getterSetter ng-model option
+  function makeTestGetterSetter(input) {
+    return function () {
+      var element = compileDirective('getterSetter', input);
+      var localValue = false;
+
+      scope.modelGetterSetter = function() {
+        return localValue;
+      };
+      scope.$apply();
+
+      expect(element.hasClass(CONST.SWITCH_OFF_CLASS)).toBeTruthy();
+      expect(element.hasClass(CONST.SWITCH_ON_CLASS)).toBeFalsy();
+
+      localValue = true;
+      scope.$apply();
+      expect(element.hasClass(CONST.SWITCH_OFF_CLASS)).toBeFalsy();
+      expect(element.hasClass(CONST.SWITCH_ON_CLASS)).toBeTruthy();
+    };
+  }
+  it('should watch updates in getterSetter', inject(makeTestGetterSetter()));
+  it('should watch updates in getterSetter', inject(makeTestGetterSetter(true)));
 
 });